GETP-193 rename: infra 패키지에서 util과 support 추출
