+++ src/main/java/es/princip/getp/domain/auth/application/VerificationService.java
@@ -1,11 +1,11 @@
 package es.princip.getp.domain.auth.application;
 
+import es.princip.getp.common.util.RandomUtil;
 import es.princip.getp.domain.auth.domain.EmailVerification;
 import es.princip.getp.domain.auth.domain.EmailVerificationRepository;
 import es.princip.getp.domain.auth.exception.IncorrectVerificationCodeException;
 import es.princip.getp.domain.auth.exception.NotFoundVerificationException;
 import es.princip.getp.domain.member.command.domain.model.Email;
-import es.princip.getp.infra.util.RandomUtil;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
