+++ src/main/java/es/princip/getp/domain/people/query/presentation/PeopleQueryController.java
@@ -3,11 +3,11 @@ package es.princip.getp.domain.people.query.presentation;
 import es.princip.getp.common.adapter.in.web.dto.ApiResponse;
 import es.princip.getp.common.adapter.in.web.dto.ApiResponse.ApiSuccessResult;
 import es.princip.getp.common.adapter.in.web.dto.PageResponse;
+import es.princip.getp.common.util.ControllerSupport;
 import es.princip.getp.domain.people.query.dao.PeopleDao;
 import es.princip.getp.domain.people.query.dto.people.CardPeopleResponse;
 import es.princip.getp.domain.people.query.dto.people.DetailPeopleResponse;
 import es.princip.getp.domain.people.query.dto.people.PublicDetailPeopleResponse;
-import es.princip.getp.infra.support.ControllerSupport;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
