+++ src/main/java/es/princip/getp/domain/project/query/dao/AppliedProjectQueryDslDao.java
@@ -1,9 +1,9 @@
 package es.princip.getp.domain.project.query.dao;
 
 import com.querydsl.jpa.impl.JPAQuery;
+import es.princip.getp.common.util.QueryDslSupport;
 import es.princip.getp.domain.project.command.domain.Project;
 import es.princip.getp.domain.project.query.dto.AppliedProjectCardResponse;
-import es.princip.getp.infra.support.QueryDslSupport;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
