+++ src/main/java/es/princip/getp/domain/project/query/dao/MyCommissionedProjectQueryDslDao.java
@@ -2,10 +2,10 @@ package es.princip.getp.domain.project.query.dao;
 
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQuery;
+import es.princip.getp.common.util.QueryDslSupport;
 import es.princip.getp.domain.project.command.domain.Project;
 import es.princip.getp.domain.project.command.domain.ProjectStatus;
 import es.princip.getp.domain.project.query.dto.MyCommissionedProjectCardResponse;
-import es.princip.getp.infra.support.QueryDslSupport;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
